+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added libcasm.sym_info.make_lattice_point_group
 - Added libcasm.configuration.Prim.lattice_point_group
+- Added libcasm.configuration.io.tools to get properties from CASM objects
 - Added libcasm.configuration.io.spglib to interface with spglib
 - Added libcasm.io.symgroup_to_dict_with_group_classification to add group classification from spglib to SymGroup.to_dict data.
 - Added libcasm.clusterography.make_local_cluster_group